search: rename package lucky -> smart
