+++ internal/search/smartsearch/generator.go
@@ -1,4 +1,4 @@
-package lucky
+package smartsearch
 
 import (
 	"strings"
