+++ internal/search/smartsearch/generator_test.go
@@ -1,4 +1,4 @@
-package lucky
+package smartsearch
 
 import (
 	"encoding/json"
