+++ internal/search/smartsearch/rules.go
@@ -1,4 +1,4 @@
-package lucky
+package smartsearch
 
 import (
 	"fmt"
