+++ internal/search/smartsearch/rules_test.go
@@ -1,4 +1,4 @@
-package lucky
+package smartsearch
 
 import (
 	"encoding/json"
