+++ internal/search/smartsearch/testdata/TestNewGenerator/rule_application#01.golden
