+++ internal/search/smartsearch/testdata/TestNewGenerator/rule_application#02.golden
