+++ internal/search/smartsearch/testdata/TestNewGenerator/rule_application.golden
