+++ internal/search/smartsearch/testdata/TestSkippedRules/do_not_apply_rules_for_type:diff.golden
