+++ internal/search/smartsearch/testdata/Test_langPatterns/lang_patterns#01.golden
