+++ internal/search/smartsearch/testdata/Test_langPatterns/lang_patterns.golden
