+++ internal/search/smartsearch/testdata/Test_patternsToCodeHostFilters/URL_patterns#05.golden
