+++ internal/search/smartsearch/testdata/Test_regexpPatterns/regexp_patterns#01.golden
