+++ internal/search/smartsearch/testdata/Test_regexpPatterns/regexp_patterns.golden
