+++ internal/search/smartsearch/testdata/Test_symbolPatterns/symbol_patterns#01.golden
