+++ internal/search/smartsearch/testdata/Test_symbolPatterns/symbol_patterns.golden
