+++ internal/search/smartsearch/testdata/Test_typePatterns/type_patterns#01.golden
