+++ internal/search/smartsearch/testdata/Test_typePatterns/type_patterns.golden
