+++ internal/search/smartsearch/testdata/Test_unorderedPatterns/AND_patterns.golden
