+++ internal/search/smartsearch/testdata/Test_unquotePatterns/unquote_patterns#01.golden
