+++ internal/search/smartsearch/testdata/Test_unquotePatterns/unquote_patterns#02.golden
