+++ internal/search/smartsearch/testdata/Test_unquotePatterns/unquote_patterns.golden
